Initial commit: MVP - minimal viable programm
This is the initial commit for `htpasswd`, a clone of the Apache utility
of the same name. The goal is to provide a compatible, portable,
stand-alone implementation when the version from Apache is not desirable
(e.g. if it is packaged together with the web server itself, as on
FreeBSD).

Known issues:

- The password is displayed during entry on stdin.
- Md5crypt is not implemented.
- Flags cannot be grouped like in classic `getopt(3)`, because they are
  implemented in plain Go uing the `flag` package.

Signed-off-by: Florian Limberger <flo@purplekraken.com>